use OneMKL nrm2 routine
Here we use nrm2 routine from OneMKL for dpnp.norm function when
possible. Using OneAPI 2024.0.0 packages the performance was not
as good as using existing functions in dpnp for calculations.
So, we keep this chnages as a back up for now.

Also, using nrm2 a batch version of it was implemented to be used when
appropriate. This need to be extended for the case when axis=2.