Use SYCL experimental extension for reduction properties (#2211)
Recent changes in DPC++ 2025.1 compiler broke DPNP compilation. It seems
the legacy DPC++ compiler behavior was affected by introducing SYCL
experimental extension for reduction properties.

Since the compilation issue occurred for a backend function which will
be removed once #2183 is merged, that PR proposes to temporary w/a the
issue and to reuse the extension for reduction properties, which seems
working.