Add `dpnp.iterable` function (#2208)
There was `__iter__` method of dpnp.ndarray implemented in scope of
#2206.

While this PR proposes to add implementation of `dpnp.iterable` function
which is intended to check if the object (including dpnp.ndarray) can be
iterated over.
The implementation leverages on NumPy call, which is based on attempt to
pass the object to `iter` function.