Align timeouts on running tests in public CI on Linux (#2267)
The timeout was set up to 12 minutes in `Conda package` workflow.
That PR proposes to align it across all GitHub workflows where tests are
running on Linux.

Sometimes the current timeout of 10 minutes might not be enough (like
[here](https://github.com/IntelPython/dpnp/actions/runs/12870503193/job/35883466675?pr=2266)
in `Test oneMKL interfaces` workflow) which caused unexpected failure or
extra rerunning attempt.