Add `mean` keyword to `dpnp.std` and `dpnp.var` (#2271)
The PR proposes to add `mean` keyword argument to `dpnp.std`,
`dpnp.var`, `dpnp.nanstd`, `dpnp.nanvar` functions and
`dpnp.ndarray.std`, `dpnp.ndarray.var` methods.

The keyword was introduced by NumPy 2.0 and intended to improve the
performance:

> Often when the standard deviation is needed the mean is also needed;
the same holds for the variance and the mean. With the current code the
mean is then calculated twice, this can be prevented if the functions
calculating the variance or the standard deviation can use a
precalculated mean

This PR implements similar improvement in dpnp code.