Add missing `message-id` to step which modifies PR comment (#2292)
`message-id` was added per #2271 to `mshick/add-pr-comment` steps in
GitHub workflows, but one of the places was missing.

The PR proposes to resolve that and to set the expected id, otherwise
the affected step couldn't find the PR comment to update it (since
`message-id: add-pr-comment` which it built by default differs).