Use `dpctl.tensor.matmul` in the backend of `dpnp.matmul` when inputs are integer (#2296)
resolves #2270 

OneMath (OneMKL) routines (`gemm`, `gemv`, `gemm_batch`) for matrix
multiplication only support floating point data types. If inputs are
integer, to use OneMath we need to upcasting them to floating point
dtypes, perform the calculation and then cast back the result to integer
dtypes which is unsafe and we may loose some information for large
integers.
In this PR, the logic for `dpnp.matmul` is updated to use
`dpctl.tensor.matmul` when result has a integer dtypes.