Return named tuple for linalg functions per python array API (#2276)
The PR proposes to return namedtuple for linalg functions: `qr`, `eig`,
`eigh`, `svd` and `slogdet`.
The namedtuple as a return type there is mandated to be compliant with
python array API.

The tests are updated to cover namedtyple as a return type of affected
function.
Also this PR requires to patch how napoleon extension from sphinx
handles docstring description of namedtuple. The handling was align with
NumPy approach.