Align signature of `dpnp.astype` with Python array API (#2318)
The PR is aimed to align signature of `dpnp.astype` function with Python
array API where only `x` and `dtype` arguments are positional-only and
other arguments are keyword-only.
Also docstrings of `dpnp.astype` function and `dpnp.ndarray,astype`
method are aligned and properly updated. And `dpnp.astype` function is
moved to array creation namespace like NumPy has it.

Additionally, docstring of `device` keyword is updated and aligned
across all the functions description to have it described in a common
way.