Update `dpnp.fft` to run on CUDA (#2332)
This PR suggests fixing the current issues with `dpnp.fft.fftn()` and
`dpnp.fft.rfftn()` on CUDA and removing the skip tests for them

The `incorrect result` issue for `dpnp.fft.fftn()` on cuda was because
preparing the input array when `batch_fft=True` could change it to `F
contiguous` array. cuFFT for correct execution requires `C contiguous`
array as input.

The issue with raising `Invalid strides` error is a bug in oneMath
([631](https://github.com/uxlfoundation/oneMath/issues/631)).
As a workaround until this issue is solved it is suggested to use swap
of the last two axes if the last dimension is 1 and there are multiple
axes.
In this case the strides inside cuFFT are calculated correctly.

Additionally updated the arguments for `test_erf` in
`skipped_tests_cuda.tbl` to skip it on cuda.