update `test_usm_type.py` (#2351)
In this PR, `test_usm_type.py` to get rid of comparison with numpy
results. The correctness of dpnp functions is separately validated in
other tests. In this file, only compute-follow-data paradigm should be
validated.
In addition, passing redundant ids into test is removed in this file and
ids for test in `test_sycl_queue.py` is simplified.