Update docstrings for elementwise functions (#2406)
The PR updates docstrings for elementwise functions to have a blank line
prior `Default` value, also to add expected dtype of input arrays where
missing, to align the dtypes with implementation and to fix some typos
which were identified.
Also a description of `include_initial` keyword is aligned for
cumulative functions.