make array Hermitian before calling complex-to-real FFT (#2444)
In this PR, input array is made Hermitian before calling complex-to-real
FFT.

Input array should be Hermitian before calling complex-to-real FFT
otherwise the behavior is undefined. Although it is user responsibility
to validate this feature,
[NumPy](https://github.com/numpy/numpy/blob/maintenance/2.2.x/numpy/fft/_pocketfft_umath.cpp#L265)
and
[CuPy](https://github.com/cupy/cupy/blob/v13.4.1/cupy/fft/_fft.py#L109)
internally make the input array Hermitian. So, it is done in dpnp for
consistency.