Explicit shape comparison for `dpnp` and `numpy` outputs (#2295)
In the test suite, the result arrays from `dpnp` and `numpy` were
compared, but their shapes were not explicitly checked for equality. As
a result, tests could pass even if the shapes differed, such as one
being `(1,)` and the other `()`.

This PR adds an explicit check to ensure that the output shapes of
`numpy` and `dpnp` match. Additionally, tests have been updated
accordingly to pass with this new check.