Update Windows version used by GitHub CI (#1841)
The PR is indented to remove a temporary pinning to "windows-2019" and
to revert changes from #1833.
Besides Windows 2019 GitHub image is going to be fully unsupported by
June 30, 2025.

The impacted workflows are moved to use "windows-2022" GH runners.
It requires updating `conda_build_config.yaml` where the version of the
compiler as well as standard libraries have to be specified.
While `{{ stdlib('c') }}` is installing into build environment as an
alignment with recipe's meta yaml from conda-forge (see
[dpctl#1868](https://github.com/IntelPython/dpctl/pull/1868) as for
reference).