Fix compilation warnings (#2517)
This PR resolves compilation warnings:
> warning: class template argument deduction for alias templates is a
C++20 extension [-Wc++20-extensions]

As for now DPNP is building with C++17 standard, and Class Template
Argument Deduction is available where only for a class template, but not
for a type alias. What support was added in C++20 standard.

Since there are currently no plans to move to the C++20 standard, the
code has been updated to conform to the C++17 standard.