Update `dpnp.isclose` with scalar-specific SYCL kernels (#2540)
This PR suggests updating `dpnp.isclose()` function adding a
scalar-specific SYCL kernels for both contiguous and stride cases to
improve performance when `rtol` and `atol` are scalars.
Also extends and updates tests for `dpnp.isclose()`

The new kernel **improves performance** by **up to 10x** compared to the
previous implementation when `rtol` and `atol` are scalars (tested on
PVC).