Resolve tests failures in rounding functions with numpy 2.0 (#2589)
There is data type mismatch in result returned for the rounding
functions, because numpy.ceil, numpy.floor, numpy.trunc, and numpy.fix
always returns floating dtype for "numpy < 2.1.0".

The dpnp behavior is aligned with "numpy >= 2.1.0", so this PR adds a
missing fixture to run the test depending on minimum required numpy
version.