Redesigned `dpnp.modf` function to be a part of `ufunc` and `vm` pybind11 extensions (#2654)
The PR updates implementation of `dpnp.modf` to move it from old legacy
backend to be part of `ufunc` and `vm` extensions.
Also `DPNPUnaryTwoOutputsFunc` class is extended with support of oneMKL
VM functions.

The tests for unary functions with 2 output arrays are separated to
`dpnp/tests/test_unary_two_outputs_ufuncs.py` file.
And the legacy implementation of `dpnp.modf` was cleaned up.